move spec override methods to spec_util
