+++ slm_lab/spec/spec_util.py
@@ -144,6 +144,50 @@ def is_aeb_compact(aeb_list):
     return aeb_compact
 
 
+def is_singleton(spec):
+    '''Check if spec uses a singleton Session'''
+    return len(spec['agent']) == 1 and len(spec['env']) == 1 and spec['body']['num'] == 1
+
+
+def override_dev_spec(spec):
+    spec['meta']['max_session'] = 1
+    spec['meta']['max_trial'] = 2
+    return spec
+
+
+def override_enjoy_spec(spec):
+    spec['meta']['max_session'] = 1
+    return spec
+
+
+def override_eval_spec(spec, num_eval_epi=100):
+    for agent_spec in spec['agent']:
+        if 'max_size' in agent_spec['memory']:
+            agent_spec['memory']['max_size'] = 100
+    for env_spec in spec['env']:
+        # evaluate by episode; offset so epi is 0 - (num_eval_epi - 1)
+        env_spec['max_tick'] = num_eval_epi - 1
+        env_spec['max_tick_unit'] = 'epi'
+    return spec
+
+
+def override_test_spec(spec):
+    for agent_spec in spec['agent']:
+        # covers episodic and timestep
+        agent_spec['algorithm']['training_frequency'] = 1
+        agent_spec['algorithm']['training_start_step'] = 1
+        agent_spec['algorithm']['training_epoch'] = 1
+        agent_spec['algorithm']['training_batch_epoch'] = 1
+    for env_spec in spec['env']:
+        env_spec['max_t'] = 20
+        env_spec['max_tick'] = 3
+        env_spec['max_tick_unit'] = 'epi'
+        env_spec['save_frequency'] = 1000
+    spec['meta']['max_session'] = 1
+    spec['meta']['max_trial'] = 2
+    return spec
+
+
 def resolve_aeb(spec):
     '''
     Resolve an experiment spec into the full list of points (coordinates) in AEB space.
