+++ test/spec/test_dist_spec.py
@@ -14,7 +14,7 @@ import pytest
 # helper method to run all tests in test_spec
 def run_trial_test_dist(spec_file, spec_name=False):
     spec = spec_util.get(spec_file, spec_name)
-    spec = util.override_test_spec(spec)
+    spec = spec_util.override_test_spec(spec)
     info_space = InfoSpace()
     info_space.tick('trial')
     spec['meta']['distributed'] = True
